% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/coord.data.frame.R
\name{coord.data.frame.R}
\alias{coord.data.frame.R}
\title{Atom coordinates to data frame}
\usage{
coord.data.frame.R(pdb, trj = NULL, sel_string)
}
\arguments{
\item{pdb}{Structure object of class pdb as read by read.pdb() funcion of bio3d lib.}

\item{trj}{(Optional) trajectory as read by read.dcd() funcion of bio3d lib.}

\item{sel_string}{Selection of atom for bio3d function atom.select(), refer to manual (es. "///71///CD/")}
}
\description{
Create a data frame with the cordinates of the selected atoms. It is based on bio3d functions.
}
\examples{
pot.pdb <- read.pdb(pot.pdbfile)
pot.trj <- read.dcd(pot.dcdfile)
K80 <- "///80///K/"
coord.data.frame (pot.pdb, pot.trj, K80)
}
\keyword{coordinates}

