\name{dataEG3movement}
\alias{dataEG3movement}
\alias{dataEG3movement}
\title{Simulated data, illustrating an outlying movement residual,  used in Foster and Bravington (2009) Section 5.1.2.}
\description{These data are stationary and have 5 chains, each with 1000 observations. The first column of this data matrix is an indicator to identify the chains. The second column is the observations. The third and final column is a constant design matrix. The outlying movement is 3 -> 4. The data were originally used to highlight the diagnostic methods, see \code{\link{diagnos.envel}}, \code{\link{diagnos}}, and \code{\link{hrplot}} for functions to perform those original tasks.}
\references{Foster, S.D. and Bravington, M.V. (2009) Graphical Diagnostics for Markov Models for Categorical Data. \emph{Journal of Computational and Graphical Statistics}, to appear.}
\keyword{misc}
