\name{dataEG4}
\alias{dataEG4}
\alias{dataEG4}
\title{Simulated non-stationary data used in Foster and Bravington (2009) Section 5.1.3.}
\description{
These data are non-stationary and have 5 chains, each with 1000 observations. The first column of this data matrix is an indicator to identify the chains. The second column is the observations. The third and fourth columns are the design matrix for the process.

These data are also used to illustrate the effect of missing covariates.
}
\references{Foster, S.D. and Bravington, M.V. (2009) Graphical Diagnostics for Markov Models for Categorical Data. \emph{Journal of Computational and Graphical Statistics}, to appear.}
\keyword{misc}
