% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/dihe.trj.R
\name{dihe.trj}
\alias{dihe.trj}
\title{Dihedral angles from Molecular Dynamics Trajectories in Matrix}
\usage{
dihe.trj(pdb, trj, dihesel, sub)
}
\arguments{
\item{pdb,}{}

\item{trj,}{}

\item{dihesel,}{}

\item{sub}{}
}
\description{
Wrap for torsion.xyz from bio3d package to compute dihedral angles and return them in a convinient matrix format
}
\examples{
pdb=protein.pdb
trj=protein.trj
L81.chi <- "//81///N,CA,CB,CG/"
ts(
  torsion.dcd(
     pdb=pdb, trj=dcd,
     L81.chi, subunits)
deltat=0.5,start=10)
}

