% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/dist.trj.R
\name{dist.trj}
\alias{dist.trj}
\title{Atom distances from Molecular Dynamics Trajectories7}
\usage{
dist.trj(pdb, trj, sel_string1, sel_string2, pdb2 = NULL, trj2 = NULL,
  pdb2sel_string1 = sel_string1, pdb2sel_string2 = sel_string2,
  ref = "mean", sub = c("A", "B", "C", "D"), subseqB = c(1:3, 4:6, 7:9,
  10:12), stride = 1, start = 1, overall = FALSE)
}
\arguments{
\item{pdb}{Structure object of class pdb as read by read.pdb() funcion of bio3d lib.}

\item{trj}{(Optional) trajectory as read by read.dcd() funcion of bio3d lib.}

\item{sel_string1}{First selection of atom for bio3d function atom.select(), refer to manual (es. "///71///CD/")}

\item{pdb2}{pdb of a reference (optional, default NULL)}

\item{trj2}{dcd of the reference (optional, default NULL)}

\item{ref}{deprecated to delete}

\item{sub}{when working on multible subunits, list of subunit, default c("A","B","C","D"),}

\item{subseqB}{sequence of subunits of the second atomselection, default c(1:3,4:6,7:9,10:12),}

\item{stride}{time interval for ts()}

\item{start}{starting time for ts()}

\item{overall}{All the data are collected as a single vector, to get an overall description of the dataset}

\item{sel_string1}{Second selection of atom for bio3d function atom.select()}

\item{pdb2sel_string1=sel_string1,}{}

\item{pdb2sel_string2=sel_string2,}{}
}
\description{
Take trajectory and pdb objects, 2 selection and return time series
of distances between the atom in selection1 and selection2 according
to sub and subseqB. It is based on bio3d functions.
}
\examples{
pdb=protein.pdb
trj=protein.trj             #trajectory
sel_string1="///80///K/"    #selection
sel_string2="///81///K/"
pdb2 <- xrays.pdb[["Code1K4C"]]         #reference
stride=Stride
start=First_frame
trj2 <- NULL
pdb2sel_string1=sel_string1
pdb2sel_string2=sel_string2

ref="mean"
sub=c("A","B","C","D")
subseqB=c(1:3,4:6,7:9,10:12)
stride=1
start=1
overall=FALSE
}
\keyword{coordinates}

