% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/filled.contour.lines.R
\name{filled.contour.lines}
\alias{filled.contour.lines}
\title{Color Filled Contour Plot with multiple options}
\usage{
filled.contour.lines(x, y = NULL, z = NULL, Nlev = NULL, scaleLev = 1,
  hlines = NULL, vlines = NULL, linessize = 1, sizeaxis = 1.6,
  contlwd = 1, ...)
}
\arguments{
\item{x}{input data, as 3d matrix or x vector}

\item{y}{vector if x is a vector (default NULL)}

\item{z}{vector if x is a vector (default NULL)}

\item{Nlev}{Number of color levels. Using Nlev=NULL (default) means Nlev=Nlev.con (as in the description)}

\item{scaleLev}{Used to compute Nlev.con, see description, (default=1)}

\item{hlines}{y intercepts of horizontal lines to draw on top of the plot}

\item{vlines}{x intercepts of vertical lines to draw on top of the plot}

\item{linessize}{line width of the optional vertical and horizontal lines (default 1)}

\item{sizeaxis}{see cex.axis (default 1.6)}

\item{contlwd}{line width of the contour lines (default 1)}

\item{...}{Additional arguments passed to filled.contour()}
}
\description{
Create Filled Contour Plot providing a number of useful option on
top of the standard functions. Number of contour levels (Nlev.con)
is automatically computed as
          Nlev.con <- (max.level - min.level)*scaleLev
}
\examples{
pmf <- read.table("pmf.dat")
names(pmf) <- c("zK12","zK3", "fes")
pmfmatrix <- acast(pmf, zK12 ~ zK3 )
pmf.x <- as.numeric(rownames(pmfmatrix) )
pmf.y <- as.numeric( colnames(pmfmatrix) )
fillcont.lines(pmf.x, pmf.y, pmfmatrix, Nlev=100,Nlev.con=3, color=palette.redtoblue, ylab=expression(z[K3]), xlab=expression(xy[K3]))
}

