% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/free.energy.R
\name{free.energy}
\alias{free.energy}
\title{Free Energy from Molecular Dynamics Distributions}
\usage{
free.energy(x, kT = 0.616, zero.min = T, norm = T)
}
\arguments{
\item{x}{normalised or unnormalised distribution of the given degree of freedom}

\item{kT}{(default 0.616)}

\item{zero.min}{set the zero at the minimum of the free energy (default T)}

\item{norm}{for debug normalisation of x can be disabled (default T)}
}
\description{
Given the distribution x of a degree of freedom of interest, the function computes the free energy at kT, according to
}
\details{
F(x) = -kT*ln( pdf(x) ) + C
}
\examples{
infile <- "test.data"
window.ll <- 20

x <- read.table(infile)
x <- ma(x, window.ll)
write.table(x,file=paste("ma",infile,sep="_"), quote=F, row.names=F, col.names=F)
}

