% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/ma.R
\name{ma}
\alias{ma}
\title{Simple moving average procedure}
\usage{
ma(x, window.length)
}
\arguments{
\item{x}{any object containing the time series (es. data.frame, ts)}

\item{window.length}{length of the windows on which averages are computed}
}
\description{
Apply a moving average procedure on time series
}
\examples{
infile <- "test.data"
window.ll <- 20

x <- read.table(infile)
x <- ma(x, window.ll)
write.table(x,file=paste("ma",infile,sep="_"), quote=F, row.names=F, col.names=F)
}

