% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/qfilled.contour.R
\name{qfilled.contour}
\alias{qfilled.contour}
\title{Quick Color Filled Contour Plot}
\usage{
qfilled.contour(x, y = NULL, z = NULL, Nlev = 10, Nlev.con = Nlev, ...)
}
\arguments{
\item{x}{input data, as 3d matrix or x vector}

\item{y}{vector if x is a vector (default NULL)}

\item{z}{vector if x is a vector (default NULL)}

\item{Nlev}{Number of color levels (default 10)}

\item{Nlev.con}{Number of contour levels (default=Nlev)}

\item{...}{Additional arguments passed to filled.contour()}
}
\description{
Quickly create Color Filled Contour Plot from three
}
\examples{
fes <- read.table("fes.dat")
qfilled.contour(fes, Nlev=100, Nlev.con=10, color=rainbow(100), xlab="d71", ylab="d64")
}

