% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rmdpalettes.R
\name{red2blue.rmcpalettes}
\alias{red2blue.rmcpalettes}
\title{Palettes of RMD package}
\usage{
red2blue.rmcpalettes(x)
}
\arguments{
\item{x}{length of the palette}
}
\description{
As per title
}
\examples{
red2blue.rmcpalettes(100)
}

