% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rmdfilled.contour.R
\name{rmdfilled.contour}
\alias{rmdfilled.contour}
\title{Custom version of filled.contour}
\usage{
rmdfilled.contour(x = seq(0, 1, length.out = nrow(z)), y = seq(0, 1,
  length.out = ncol(z)), z, xlim = range(x, finite = TRUE), ylim = range(y,
  finite = TRUE), zlim = range(z, finite = TRUE), levels = pretty(zlim,
  nlevels), nlevels = 20, color.palette = cm.colors,
  col = color.palette(length(levels) - 1), plot.title, plot.axes, key.title,
  key.axes, asp = NA, xaxs = "i", yaxs = "i", las = 1, axes = TRUE,
  frame.plot = axes, border = F, cexlegend = 1, ...)
}
\arguments{
\item{border}{border of the legend (default F)}

\item{cexlegend}{legend font size}
}
\description{
Version of filled.contour() used by the RMD package. Additional parameter provided
}
\examples{
see ?filled.contour()
}

