% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rmdsmooth.spline.R
\name{rmdsmooth.spline}
\alias{rmdsmooth.spline}
\title{Quick smoothing with spline}
\usage{
rmdsmooth.spline(x, y = NULL, nx, ...)
}
\arguments{
\item{x}{a vector of  a two column matrix for smooth.spline()}

\item{y}{needed only if x is a vector (default NULL)}

\item{nx}{number of x points for the smoothed curve, supplied to predict()}

\item{...}{additional arguments passed to smooth.spline}
}
\description{
Perform the fit to a cubic smoothing spline and the prediction of the smoothed curved in a single step.
}
\examples{
x<-read.table("data")
plot(rmdsmooth.spline(x, 80))
}

